update
On branch master
Changes to be committed:
modified:   AddToDatabase.sh
modified:   Delete.sh
new file:   DeleteMapping.sh
new file:   GetAllDocs.sh
new file:   GetMapping.sh
new file:   MakeMapping.sh
modified:   SearchIndexType.sh

+++ Delete.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
+
+if [ $# -ne 1 ]; then
+  echo '$1 = doc id'
+  exit
+fi
+
 ## add document to database
 index="workouts"
 type="workout"
-docID="_search"
+docID=$1
 
 ## add document
 curl -X DELETE "http://127.0.0.1:9200/${index}/${type}/${docID}"
